把~say指令加上管理員限制器

+++ cogs/cmd.py
@@ -11,8 +11,11 @@ class cmd(Cog_Extension):
  #刪除所傳的訊息並讓機器人覆誦
   @commands.command()
   async def say(self,ctx,msg):
-      await ctx.message.delete()
-      await ctx.send(msg)
+      if ctx.author.guild_permissions.administrator:
+          await ctx.message.delete()
+          await ctx.send(msg)
+      else:
+          await ctx.send("你沒有管理者權限用來執行這個指令")
 
  #刪除所選數量的訊息
   @commands.command()
